개선: 코드 리뷰 피드백 반영 - SDK 보호 키워드 단일화 및 패턴 정교화
- MessageContainsSdkKeyword 헬퍼 추출 — IsKnownNonSdkMessage 가드와 DetermineErrorSource 분류가 AitKeywords 전체(AITNpmRunner/AITConvertCore 등 포함)를 단일 source로 공유
- SdkMessagePatterns에서 AitKeywords와 중복되던 "AppsInToss"/"apps-in-toss" 제거 (drift 방지)
- "Fields serialized in" 패턴을 "Fields serialized in [Assembly-CSharp]"로 좁혀 SDK 어셈블리 직렬화 경고 false positive 방지
- ".meta) exists but its folder" → "exists but its folder"로 변경, Unity 버전별 괄호 유무에 견고
- IsKnownNonSdkMessageTests를 별도 파일로 분리, AitKeywords 전체 키워드에 대한 negative 테스트 추가 (AITNpmRunner/AITConvertCore/AITPackageBuilder/AITNodeJS/AIT:/ait-build)
- NonSdkMessagePatterns 위에 새 노이즈 패턴 추가 워크플로우 코멘트 추가
